Add config.yaml file and specify Python interpreter path
This commit adds a new config.yaml file to the repository. The config file specifies the path to the Python interpreter to be used in the workflow. Additionally, it defines sample data paths and sets a prior mutation rate.

Note: Due to issues with installing  via Bioconda, the Python interpreter path is explicitly defined to ensure compatibility with the workflow.